refactor!: Factor-out rossum_models from rossum_api
The models are discovered as a plugin that can be provided by a 3rd
party.
