--- rossum_api/__init__.py
@@ -1,12 +1,0 @@
-from __future__ import annotations
-
-from rossum_api.api_client import APIClientError
-from rossum_api.elis_api_client import ElisAPIClient, ExportFileFormats
-from rossum_api.elis_api_client_sync import ElisAPIClientSync
-
-__all__ = (
-    "APIClientError",
-    "ElisAPIClient",
-    "ElisAPIClientSync",
-    "ExportFileFormats",
-)
