+++ rossum_sdk/rossum_api/elis_api_client_sync.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import asyncio
 import typing
 
-from rossum_api import ElisAPIClient
+from rossum_sdk.rossum_api import ElisAPIClient
 
 if typing.TYPE_CHECKING:
     import pathlib
@@ -23,18 +23,18 @@ if typing.TYPE_CHECKING:
 
     import httpx
 
-    from rossum_api import ExportFileFormats
-    from rossum_api.api_client import APIClient
-    from rossum_api.models.annotation import Annotation
-    from rossum_api.models.connector import Connector
-    from rossum_api.models.hook import Hook
-    from rossum_api.models.inbox import Inbox
-    from rossum_api.models.organization import Organization
-    from rossum_api.models.queue import Queue
-    from rossum_api.models.schema import Schema
-    from rossum_api.models.user import User
-    from rossum_api.models.user_role import UserRole
-    from rossum_api.models.workspace import Workspace
+    from rossum_sdk.rossum_api import ExportFileFormats
+    from rossum_sdk.rossum_api.api_client import APIClient
+    from rossum_sdk.rossum_api.models.annotation import Annotation
+    from rossum_sdk.rossum_api.models.connector import Connector
+    from rossum_sdk.rossum_api.models.hook import Hook
+    from rossum_sdk.rossum_api.models.inbox import Inbox
+    from rossum_sdk.rossum_api.models.organization import Organization
+    from rossum_sdk.rossum_api.models.queue import Queue
+    from rossum_sdk.rossum_api.models.schema import Schema
+    from rossum_sdk.rossum_api.models.user import User
+    from rossum_sdk.rossum_api.models.user_role import UserRole
+    from rossum_sdk.rossum_api.models.workspace import Workspace
 
     T = TypeVar("T")
