+++ rossum_sdk/rossum_models/connector.py
