+++ tests/rossum_api/conftest.py
@@ -5,8 +5,8 @@ import pytest
 import pytest_asyncio
 from mock import MagicMock
 
-from rossum_api import ElisAPIClient, ElisAPIClientSync
-from rossum_api.api_client import APIClient
+from rossum_sdk.rossum_api import ElisAPIClient, ElisAPIClientSync
+from rossum_sdk.rossum_api.api_client import APIClient
 
 
 @pytest.fixture
