+++ tests/rossum_api/elis_api_client/test_annotations.py
@@ -3,10 +3,13 @@ from __future__ import annotations
 import pytest
 from mock.mock import MagicMock, patch
 
-from rossum_api.models.annotation import Annotation
-from rossum_api.models.automation_blocker import AutomationBlocker, AutomationBlockerContent
-from rossum_api.models.document import Document
-from rossum_api.models.user import User
+from rossum_sdk.rossum_api.models.annotation import Annotation
+from rossum_sdk.rossum_api.models.automation_blocker import (
+    AutomationBlocker,
+    AutomationBlockerContent,
+)
+from rossum_sdk.rossum_api.models.document import Document
+from rossum_sdk.rossum_api.models.user import User
 
 
 @pytest.fixture
@@ -296,8 +299,7 @@ class TestAnnotationsSync:
             annotation = Annotation(**dummy_annotation_with_sideloads)
             modifier = User(**annotation.modifier)
             annotation.modifier = modifier
-            automation_blocker = Document(**annotation.document)
-            annotation.document = automation_blocker
+            annotation.document = Document(**annotation.document)
             automation_blocker = AutomationBlocker(**annotation.automation_blocker)
             automation_blocker.content = [
                 AutomationBlockerContent(**content) for content in automation_blocker.content
