+++ tests/rossum_api/elis_api_client/test_workspaces.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from rossum_api.models.workspace import Workspace
+from rossum_sdk.rossum_api.models.workspace import Workspace
 
 
 @pytest.fixture
