+++ tests/rossum_api/test_api_client.py
@@ -18,7 +18,7 @@ import mock
 import pytest
 import pytest_httpx
 
-from rossum_api.api_client import APIClient, APIClientError
+from rossum_sdk.rossum_api.api_client import APIClient, APIClientError
 
 WORKSPACES = [
     {
